Dry up code by moving methods to statsigHelpers.js
