+++ apps/src/metrics/StatsigReporter.js
@@ -1,6 +1,5 @@
 import {StatsigClient} from '@statsig/js-client';
 import {runStatsigAutoCapture} from '@statsig/web-analytics';
-import cookies from 'js-cookie';
 
 import logToCloud from '@cdo/apps/logToCloud';
 import experiments from '@cdo/apps/util/experiments';
@@ -10,20 +9,26 @@ import {
   getEnvironment,
   isProductionEnvironment,
   isDevelopmentEnvironment,
-  createUuid,
 } from '../utils';
+
+import {
+  getUserID,
+  getUserType,
+  findOrCreateStableId,
+  formatUserId,
+} from './StatsigHelpers';
 
 // A flag that can be toggled to send events regardless of environment
 const ALWAYS_SEND = false;
 const NO_EVENT_NAME = 'NO_VALID_EVENT_NAME_LOG_ERROR';
-const STABLE_ID_KEY = 'statsig_stable_id';
+// const STABLE_ID_KEY = 'statsig_stable_id';
 
 class StatsigReporter {
   constructor() {
     // stable_id is set as a cookie in application_controller.rb. However in a
     // the rare case we are running outside of the application layout,
     // set stable_id as a cookie here if it doesn't exist.
-    this.stable_id = this.findOrCreateStableId();
+    this.stable_id = findOrCreateStableId();
     this.log(`Statsig Stable ID: ${this.stable_id}`);
     let user = {
       custom: {
@@ -32,14 +37,12 @@ class StatsigReporter {
       },
       customIDs: {stableID: this.stable_id},
     };
-    const user_id_element = document.querySelector('script[data-user-id]');
-    const user_id = user_id_element ? user_id_element.dataset.userId : null;
-    const user_type_element = document.querySelector('script[data-user-type');
-    const user_type = user_type_element
-      ? user_type_element.dataset.userType
-      : null;
+
+    const user_id = getUserID();
+    const user_type = getUserType();
+
     if (user_id) {
-      user.userID = this.formatUserId(user_id);
+      user.userID = formatUserId(user_id);
       user.custom.userType = user_type;
     }
     this.user = user;
@@ -86,7 +89,7 @@ class StatsigReporter {
     enabledExperiments,
     educatorRole,
   }) {
-    const formattedUserId = this.formatUserId(userId);
+    const formattedUserId = formatUserId(userId);
     const user = {
       userID: formattedUserId,
       custom: {
@@ -144,30 +147,6 @@ class StatsigReporter {
     return (
       this.statsigClient.getExperiment(name).value[parameter] ?? defaultValue
     );
-  }
-
-  formatUserId(userId) {
-    const userIdString = userId.toString() || 'none';
-    if (!userId) {
-      return userIdString;
-    }
-    if (isProductionEnvironment()) {
-      return userIdString.padStart(5, '0');
-    } else {
-      const environment = getEnvironment();
-      return `${environment}-${userIdString}`;
-    }
-  }
-
-  findOrCreateStableId() {
-    let stableId = cookies.get(STABLE_ID_KEY);
-    if (!stableId) {
-      stableId = createUuid();
-      cookies.set(STABLE_ID_KEY, stableId, {
-        path: '/',
-      });
-    }
-    return stableId;
   }
 
   /**
