+++ apps/src/metrics/StatsigSessionReplay.js
@@ -1,30 +1,30 @@
 import {StatsigClient} from '@statsig/js-client';
 import {runStatsigSessionReplay} from '@statsig/session-replay';
 import {runStatsigAutoCapture} from '@statsig/web-analytics';
-import cookies from 'js-cookie';
 
-import {getEnvironment, isProductionEnvironment, createUuid} from '../utils';
-
-const STABLE_ID_KEY = 'statsig_stable_id';
+import {
+  getUserID,
+  getUserType,
+  findOrCreateStableId,
+  formatUserId,
+} from './StatsigHelpers';
 
 class StatsigSessionReplay {
   constructor() {
     // stable_id is set as a cookie in application_controller.rb. However in a
     // the rare case we are running outside of the application layout,
     // set stable_id as a cookie here if it doesn't exist.
-    this.stable_id = this.findOrCreateStableId();
+    this.stable_id = findOrCreateStableId();
     let user = {
       customIDs: {stableID: this.stable_id},
       custom: {},
     };
-    const user_id_element = document.querySelector('script[data-user-id]');
-    const user_id = user_id_element ? user_id_element.dataset.userId : null;
-    const user_type_element = document.querySelector('script[data-user-type');
-    const user_type = user_type_element
-      ? user_type_element.dataset.userType
-      : null;
+
+    const user_id = getUserID();
+    const user_type = getUserType();
+
     if (user_id) {
-      user.userID = this.formatUserId(user_id);
+      user.userID = formatUserId(user_id);
       user.custom.userType = user_type;
     }
     this.user = user;
@@ -55,30 +55,6 @@ class StatsigSessionReplay {
 
     this.statsigClient.shutdown();
     this.statsigClient = null;
-  }
-
-  formatUserId(userId) {
-    const userIdString = userId.toString() || 'none';
-    if (!userId) {
-      return userIdString;
-    }
-    if (isProductionEnvironment()) {
-      return userIdString.padStart(5, '0');
-    } else {
-      const environment = getEnvironment();
-      return `${environment}-${userIdString}`;
-    }
-  }
-
-  findOrCreateStableId() {
-    let stableId = cookies.get(STABLE_ID_KEY);
-    if (!stableId) {
-      stableId = createUuid();
-      cookies.set(STABLE_ID_KEY, stableId, {
-        path: '/',
-      });
-    }
-    return stableId;
   }
 }
